requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
boto3==1.34.62
botocore==1.34.62
CacheControl==0.14.0
cachetools==5.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.39.0
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.4
django-allauth==0.61.1
django-apscheduler==0.6.2
django-celery-beat==2.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-environ==0.11.2
django-extra-fields==3.0.2
django-filter==24.2
django-ipware==6.0.4
django-phonenumber-field==7.3.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.4.0
django-timezone-field==6.1.0
django-user-agents==0.4.0
django_db==0.0.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dynaconf==3.2.4
filelock==3.13.1
firebase-admin==6.5.0
frozenlist==1.4.1
geoip2==4.8.0
google-api-core==2.18.0
google-api-python-client==2.65.0
google-auth==2.29.0
google-auth-httplib2==0.1.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.56.4
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.21.0
hvac==2.1.0
idna==3.6
inflection==0.5.1
install==1.3.5
Jinja2==3.1.3
jmespath==1.0.1
knox==0.1.14
kombu==5.3.5
loguru==0.7.2
MarkupSafe==2.1.5
maxminddb==2.5.2
msgpack==1.0.8
multidict==6.0.5
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.21.9
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.0.9
pypaystack==1.24
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python-ipware==2.0.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
s3transfer==0.10.0
sendgrid==6.11.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
timedelta==2020.12.3
typing_extensions==4.10.0
tzdata==2023.4
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.7
user-agents==2.2.0
validators==0.22.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
whitenoise==6.6.0
yarl==1.9.4
